Move earth/sd/* into earth/dev_disk.c
